+++ src/app/app.component.spec.ts
@@ -1,16 +1,28 @@
 import { TestBed } from '@angular/core/testing';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppComponent } from './app.component';
+import { TranslateModule } from '@ngx-translate/core';
+import { MenuComponent } from '@shared/components/menu/menu.component';
+import { LoggerComponent } from '@shared/components/logger/logger.component';
+import { FooterComponent } from '@shared/components/footer/footer.component';
+import { socialAuthServiceConfigProvider } from '@shared/constants/social.constants';
 
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
       imports: [
+        TranslateModule.forRoot(),
         RouterTestingModule
       ],
       declarations: [
-        AppComponent
+        AppComponent,
+        MenuComponent,
+        LoggerComponent,
+        FooterComponent,
       ],
+      providers: [
+        socialAuthServiceConfigProvider
+      ]
     }).compileComponents();
   });
 
@@ -24,12 +36,5 @@ describe('AppComponent', () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.componentInstance;
     expect(app.title).toEqual('carrito');
-  });
-
-  it('should render title', () => {
-    const fixture = TestBed.createComponent(AppComponent);
-    fixture.detectChanges();
-    const compiled = fixture.nativeElement as HTMLElement;
-    expect(compiled.querySelector('h1')?.textContent).toContain('Hello, carrito');
   });
 });