+++ src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { APP_INITIALIZER, Injector, NgModule } from '@angular/core';
 import { BrowserModule, provideClientHydration } from '@angular/platform-browser';
-
+import { GoogleSigninButtonModule, SocialLoginModule } from '@abacritt/angularx-social-login';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
@@ -8,6 +8,10 @@ import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TranslateLoader, TranslateModule, TranslateService } from '@ngx-translate/core';
 import { SharedModule } from '@shared/shared.module';
+import { socialAuthServiceConfigProvider } from '@shared/constants/social.constants';
+import { MenuComponent } from '@shared/components/menu/menu.component';
+import { FooterComponent } from '@shared/components/footer/footer.component';
+import { LoggerComponent } from '@shared/components/logger/logger.component';
 
 export const createTranslateLoader = (http: HttpClient) => {
   return new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -23,6 +27,9 @@ function appInitializerFactory(translateService: TranslateService) {
 @NgModule({
   declarations: [
     AppComponent,
+    MenuComponent,
+    FooterComponent,
+    LoggerComponent,
   ],
   imports: [
     BrowserModule,
@@ -36,7 +43,10 @@ function appInitializerFactory(translateService: TranslateService) {
           useFactory: (createTranslateLoader),
           deps: [HttpClient]
       }
-  })
+    }),
+    SocialLoginModule,
+    GoogleSigninButtonModule,
+
   ],
   providers: [
     provideClientHydration(),
@@ -45,7 +55,8 @@ function appInitializerFactory(translateService: TranslateService) {
       useFactory: appInitializerFactory,
       deps: [TranslateService, Injector],
       multi: true
-    }
+    },
+    socialAuthServiceConfigProvider
   ],
   bootstrap: [AppComponent]
 })