+++ src/app/shared/constants/routes.constants.ts
@@ -3,5 +3,17 @@ export const ROUTES = {
   HOME: {
     path: '',
     filePath: './modules/+home/home.module',
-  }
+  },
+  LISTS: {
+    path: 'lists',
+    filePath: './modules/+lists/lists.module',
+  },
+  RECIPES: {
+    path: 'recipes',
+    filePath: './modules/+recipes/recipes.module',
+  },
+  COMUNITY: {
+    path: 'comunity',
+    filePath: './modules/+comunity/comunity.module',
+  },
 };